requirements.txt
click==6.6
ecdsa==0.13
Flask>=0.11.1
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Naked==0.1.31
nexmo==2.0.0
ordereddict==1.1
pycrypto==2.6.1
python-jose==1.2.0
PyYAML==3.12
requests==2.10.0
shellescape==3.4.1
six==1.10.0
Werkzeug==0.11.10
python-dotenv==0.7.1
voice>=voice-0.1.0
Flask==0.11.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-notifyAll==1.0.3
google-api-python-client==1.6.7
nexmo==2.0.0
picamera==1.13
